JHA: Updated project to be blank slate.
