+++ jha/src/store/index.js
@@ -1,15 +1,13 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import app from './modules/app';
-import payPatientForm from './modules/pay-patient-form';
-import payPractitionerForm from './modules/pay-practitioner-form';
+import appModule from './modules/app-module';
+import formAModule from './modules/form-a-module';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
   modules: {
-    app,
-    payPatientForm,
-    payPractitionerForm,
+    appModule,
+    formAModule,
   }
 });
