+++ jha/src/views/form-a/FormPage.vue
@@ -1,0 +1,120 @@
+<template>
+  <div>
+    <PageContent>
+      <div class="container pt-3 pt-sm-5 mb-3">
+        <div class="row align-items-end">
+          <div class="col-md-7">
+            <h1>Form</h1>
+          </div>
+          <div class="col-md-5">
+            <p class="text-right"><span class="required-asterisk">*</span> Required Information</p>
+          </div>
+        </div>
+        <hr class="mt-0"/>
+        <p>Form here.</p>
+      </div>
+    </PageContent>
+    <ContinueBar @continue="validateFields()" />
+  </div>
+</template>
+
+<script>
+import pageStateService from '@/services/page-state-service';
+import {
+  formARoutes,
+  isPastPath,
+} from '@/router/routes';
+import {
+  scrollTo,
+  scrollToError,
+  getTopScrollPosition
+} from '@/helpers/scroll';
+import {
+  getConvertedPath,
+} from '@/helpers/url';
+import ContinueBar from '@/components/ContinueBar.vue';
+import PageContent from '@/components/PageContent.vue';
+import {
+  MODULE_NAME as formModule,
+  RESET_FORM,
+} from '@/store/modules/form-a-module';
+import logService from '@/services/log-service';
+
+export default {
+  name: 'MainFormPage',
+  components: {
+    ContinueBar,
+    PageContent,
+  },
+  data: () => {
+    return {
+      isPageLoaded: false,
+    };
+  },
+  created() {
+    setTimeout(() => {
+      this.isPageLoaded = true;
+    }, 0);
+
+    logService.logNavigation(
+      this.$store.state.formAModule.applicationUuid,
+      formARoutes.FORM_PAGE.path,
+      formARoutes.FORM_PAGE.title
+    );
+  },
+  validations() {
+    const validations = {};
+    return validations;
+  },
+  methods: {
+    validateFields() {
+      this.$v.$touch()
+      if (this.$v.$invalid) {
+        scrollToError();
+        return;
+      }
+
+      this.navigateToNextPage();
+    },
+    navigateToNextPage() {
+      // Navigate to next path.
+      const toPath = getConvertedPath(
+        this.$router.currentRoute.path,
+        formARoutes.REVIEW_PAGE.path
+      );
+      pageStateService.setPageComplete(toPath);
+      pageStateService.visitPage(toPath);
+      this.$router.push(toPath);
+      scrollTo(0);
+    },
+  },
+  computed: {},
+  // Required in order to block back navigation.
+  beforeRouteLeave(to, from, next) {
+    pageStateService.setPageIncomplete(from.path);
+    if (to.path === formARoutes.HOME_PAGE.path) {
+      this.$store.dispatch(formModule + '/' + RESET_FORM);
+      next();
+    } else if ((pageStateService.isPageComplete(to.path)) || isPastPath(to.path, from.path)) {
+      next();
+    } else {
+      // Navigate to self.
+      const topScrollPosition = getTopScrollPosition();
+      const toPath = getConvertedPath(
+        this.$router.currentRoute.path,
+        formARoutes.FORM_PAGE.path
+      );
+      next({
+        path: toPath,
+        replace: true
+      });
+      setTimeout(() => {
+        scrollTo(topScrollPosition);
+      }, 0);
+    }
+  }
+}
+</script>
+
+<style scoped>
+</style>
