+++ jha/src/views/form-a/SubmissionErrorPage.vue
@@ -23,19 +23,16 @@
 </template>
 
 <script>
-import PageContent from '../../components/PageContent.vue';
-import pageStateService from '../../services/page-state-service';
-import {
-  payPatientRoutes,
-  payPatientCSRRoutes
-} from '../../router/routes';
+import PageContent from '@/components/PageContent.vue';
+import pageStateService from '@/services/page-state-service';
+import { formARoutes } from '@/router/routes';
 import {
   MODULE_NAME as formModule,
   RESET_FORM
-} from '../../store/modules/pay-patient-form';
-import { scrollTo } from '../../helpers/scroll';
+} from '@/store/modules/form-a-module';
+import { scrollTo } from '@/helpers/scroll';
 import { getConvertedPath } from '@/helpers/url';
-import logService from '../../services/log-service';
+import logService from '@/services/log-service';
 
 export default {
   name: 'SubmissionErrorPage',
@@ -44,22 +41,21 @@ export default {
   },
   created() {
     logService.logNavigation(
-      this.$store.state.payPatientForm.applicationUuid,
-      payPatientRoutes.SUBMISSION_ERROR_PAGE.path,
-      payPatientRoutes.SUBMISSION_ERROR_PAGE.title
+      this.$store.state.formAModule.applicationUuid,
+      formARoutes.SUBMISSION_ERROR_PAGE.path,
+      formARoutes.SUBMISSION_ERROR_PAGE.title
     );
   },
   // Required in order to block back navigation.
   beforeRouteLeave(to, from, next) {
     pageStateService.setPageIncomplete(from.path);
     this.$store.dispatch(formModule + '/' + RESET_FORM);
-    if (to.path === payPatientRoutes.HOME_PAGE.path
-      || to.path === payPatientCSRRoutes.HOME_PAGE.path) {
+    if (to.path === formARoutes.HOME_PAGE.path) {
       next();
     } else {
       const toPath = getConvertedPath(
         this.$router.currentRoute.path,
-        payPatientRoutes.HOME_PAGE.path
+        formARoutes.HOME_PAGE.path
       )
       next({ path: toPath });
     }
